+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to the `swagger-explorer` extension will be documented in th
 
 ## [0.1.0] Future
 
+* [x] update for new vscode versions
 * [ ] [Refresh per source](https://github.com/dardino/vscode-swagger-explorer/issues/11)
 * [ ] [Swagger JSON validation](https://github.com/dardino/vscode-swagger-explorer/issues/12)